feature (#110) : update syntax in `Main`, refactor type return in `Statistic` and add a test in `StatisticTest`

+++ src/main/java/fr/univ_cotedazur/polytech/si3/team_c/citadels/Statistic.java
@@ -19,15 +19,15 @@ public class Statistic {
         equality++;
     }
 
-    public double getWin() {
+    public int getWin() {
         return win;
     }
 
-    public double getLoose() {
+    public int getLoose() {
         return loose;
     }
 
-    public double getEquality() {
+    public int getEquality() {
         return equality;
     }
